+++ lib/bluetooth/read.js
@@ -1,4 +1,4 @@
-const shared = require('./shared.js');
+const shared = require('../shared.js');
 const Promise = require('bluebird');
 
 /**
