+++ lib/exec.js
@@ -1,9 +1,9 @@
-const awoxConnect = require('./bluetooth.connect.js');
-const awoxDiscoverServices = require('./bluetooth.discoverServices.js');
-const awoxDiscoverCharacteristics = require('./bluetooth.discoverCharacteristics.js');
-const awoxSend = require('./bluetooth.send.js');
-const awoxScan = require('./bluetooth.scan.js');
-const generateCommand = require('./generateCommand.js');
+const awoxConnect = require('./bluetooth/connect.js');
+const awoxDiscoverServices = require('./bluetooth/discoverServices.js');
+const awoxDiscoverCharacteristics = require('./bluetooth/discoverCharacteristics.js');
+const awoxSend = require('./bluetooth/send.js');
+const awoxScan = require('./bluetooth/scan.js');
+const generateCommand = require('./default/generateCommand.js');
 
 const meshExec = require('./mesh/exec.js');
 const meshShared = require('./mesh/shared.js');
