+++ lib/install.js
@@ -1,8 +1,8 @@
-const awoxScan = require('./bluetooth.scan.js');
-const awoxConnect = require('./bluetooth.connect.js');
-const awoxDiscoverServices = require('./bluetooth.discoverServices.js');
-const awoxDiscoverCharacteristics = require('./bluetooth.discoverCharacteristics.js');
-const awoxRead = require('./bluetooth.read.js');
+const awoxScan = require('./bluetooth/scan.js');
+const awoxConnect = require('./bluetooth/connect.js');
+const awoxDiscoverServices = require('./bluetooth/discoverServices.js');
+const awoxDiscoverCharacteristics = require('./bluetooth/discoverCharacteristics.js');
+const awoxRead = require('./bluetooth/read.js');
 const managePeripheral = require('./managePeripheral.js');
 
 const Promise = require('bluebird');
