+++ lib/managePeripheral.js
@@ -1,5 +1,5 @@
-const Promise = require('bluebird');
 const shared = require('./shared.js');
+const Promise = require('bluebird');
 
 const managedDeviceNamePrefix = ['sml'];
