+++ lib/mesh/changeAuthentication.js
@@ -1,7 +1,7 @@
 const Promise = require('bluebird');
 
-const awoxSend = require('../bluetooth.send.js');
-const awoxRead = require('../bluetooth.read.js');
+const awoxSend = require('../bluetooth/send.js');
+const awoxRead = require('../bluetooth/read.js');
 const meshUtils = require('./commandUtils.js');
 const meshShared = require('./shared.js');
