+++ lib/mesh/commandUtils.js
@@ -119,7 +119,7 @@ module.exports = {
     payload.fill(command, 2, 3);
     payload.fill(0x60, 3, 4);
     payload.fill(0x01, 4, 5);
-    payload.fill(data, 5, 5 + Math.min(data.length, 15));
+    payload.fill(data, 5, 5 + data.length);
 
     // Compute checksum
     const checksum = this.checksum(sessionKey, nonce, payload);
