+++ lib/mesh/exec.js
@@ -1,6 +1,6 @@
-const awoxSend = require('../bluetooth.send.js');
+const awoxSend = require('../bluetooth/send.js');
 const generateCommand = require('./generateCommand.js');
-const meshShared = require('./shared.js');
+const meshShared = require('../shared.js');
 const meshAuthenticate =require('./authenticate.js');
 
 module.exports = function (peripheral, characteristics, type, value) {
