+++ test/unit/lib/mesh/authenticate.test.js
@@ -52,8 +52,8 @@ const awoxReadMock = function (peripheral, characteristic) {
 
 const authenticate = proxyquire('../../../../lib/mesh/authenticate.js', {
   './commandUtils.js': meshUtilsMock,
-  '../bluetooth.send.js': awoxSendMock,
-  '../bluetooth.read.js': awoxReadMock
+  '../bluetooth/send.js': awoxSendMock,
+  '../bluetooth/read.js': awoxReadMock
 });
 
 describe('Gladys authenticate AwoX Mesh network', function () {
